zoom + some canvas changes
Added zooming capability by resizing the background image on the
canvas. This is a quick-fix, because free-hand/annotations added to the
canvas do not scale with the image. We need to actually draw the image
on the canvas itself, and resize the canvas. Drawing the image on the
canvas also enables us to save the entire image (image + free hand
drawings). That's next on my TODO list :)