More work on the zoom function
Added a second canvas for the zoomed in portion of the image. Select
the radio "zoom", click and drag over the area you want to zoom in on,
and you'll be able to draw on a zoom-ed in version while your marks are
replicated on the original. Canvas saves through view changes.

TODO: lots of magic numbers to clean up; pain hierarchy, tagging vs.
canvas functionality etc.